refactor: page 관련 request와 response 클래스를 common.page로 이동
